Split knowledge base database modules
